refactor: purchaseAmount 코드 수정 및 테스트 작성
- purchase() 메서드 추가
- purchase() 메서드 관련 테스트 코드 작성
- 패키지 생성으로 테스트코드 패키지 변경
