+++ src/test/java/lotto/domain/PurchaseAmountTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 
 class PurchaseAmountTest {
 
@@ -21,5 +23,17 @@ class PurchaseAmountTest {
 		assertThatThrownBy(() -> new PurchaseAmount("999"))
 			.isInstanceOf(IllegalArgumentException.class)
 			.hasMessage("[ERROR] 구매금액은 최소 1000원 이상입니다.");
+	}
+
+	@ParameterizedTest
+	@CsvSource(value = {"4501|4", "2222|2", "2000|2", "14333|14", "7899|7"}, delimiter = '|')
+	@DisplayName("로또 구매 시 수량 확인 테스트")
+	public void PurchaseAmountPurchaseTest(String inputValue, int expectedValue) {
+		//given
+		PurchaseAmount purchaseAmount = new PurchaseAmount(inputValue);
+		//when
+		int purchaseQuantity = purchaseAmount.purchase();
+		//then
+		assertThat(purchaseQuantity).isEqualTo(expectedValue);
 	}
 }
